Configure Github Actions Workflow to run Nx affected commands and TypeOrm migrations
