+++ apps/backend/src/config/config.service.ts
@@ -1,13 +1,9 @@
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-import { InitialSchema1619147036371 } from '../migrations/1619147036371-Initial_Schema';
-import { Portfolio } from '../models/portfolio.entity';
-import { PortfolioRating } from '../models/portfolioRating.entity';
-import { PortfolioStock } from '../models/portfolioStock.entity';
-import { StockRating } from '../models/stockRating.entity';
-import { UserAccount } from '../models/userAccount.entity';
+import typeOrmConfig = require('./ormconfig');
 
 require('dotenv').config();
 
+/** Used to connect the application to the database. */
 export class ConfigService {
   constructor(private env: { [k: string]: string | undefined }) {}
 
@@ -25,32 +21,9 @@ export class ConfigService {
     return this;
   }
 
-  public getPort() {
-    return this.getValue('PORT', true);
-  }
-
-  public isProduction() {
-    const mode = this.getValue('MODE', false);
-    return mode != 'DEV';
-  }
-
   public getTypeOrmConfig(): TypeOrmModuleOptions {
     console.log('CONNECTING TO DATABASE!');
-    return {
-      type: 'postgres',
-      url: this.getValue('DATABASE_URL'),
-      ssl: this.isProduction(),
-      logging: true,
-      entities: [Portfolio, PortfolioRating, PortfolioStock, StockRating, UserAccount],
-      // entities: [__dirname + '/models/*.entity{.ts,.js}'],
-      // migrationsRun: true,
-      migrationsTableName: 'migration',
-      // migrations: [__dirname + '/migrations/*.ts'],
-      migrations: [InitialSchema1619147036371],
-      cli: {
-        migrationsDir: 'src/migration',
-      },
-    };
+    return typeOrmConfig;
   }
 }
