--- server.ts
@@ -1,3 +1,0 @@
-import 'zone.js/dist/zone-node';
-import './server/main';
-export * from './src/main.server';
